reafactor: divide inspector in components
