+++ components/devtools/actors/inspector.rs
@@ -9,19 +9,44 @@
 use std::cell::RefCell;
 use std::net::TcpStream;
 
-use base::id::PipelineId;
-use devtools_traits::DevtoolScriptControlMsg::{
-    GetChildren, GetDocumentElement, GetLayout, GetRootNode, ModifyAttribute,
-};
-use devtools_traits::{ComputedNodeLayout, DevtoolScriptControlMsg, NodeInfo};
+use devtools_traits::DevtoolScriptControlMsg;
+use devtools_traits::DevtoolScriptControlMsg::GetRootNode;
 use ipc_channel::ipc::{self, IpcSender};
 use serde::Serialize;
 use serde_json::{self, Map, Value};
 
 use crate::actor::{Actor, ActorMessageStatus, ActorRegistry};
 use crate::actors::browsing_context::BrowsingContextActor;
+use crate::actors::inspector::highlighter::{HighlighterActor, HighlighterMsg};
+use crate::actors::inspector::node::NodeInfoToProtocol;
+use crate::actors::inspector::page_style::{PageStyleActor, PageStyleMsg};
+use crate::actors::inspector::walker::{WalkerActor, WalkerMsg};
 use crate::protocol::JsonPacketStream;
-use crate::{EmptyReplyMsg, StreamId};
+use crate::StreamId;
+
+mod highlighter;
+mod node;
+mod page_style;
+mod walker;
+
+#[derive(Serialize)]
+#[serde(rename_all = "camelCase")]
+struct GetPageStyleReply {
+    from: String,
+    page_style: PageStyleMsg,
+}
+
+#[derive(Serialize)]
+struct GetWalkerReply {
+    from: String,
+    walker: WalkerMsg,
+}
+
+#[derive(Serialize)]
+struct GetHighlighterReply {
+    from: String,
+    highligter: HighlighterMsg, // sic.
+}
 
 pub struct InspectorActor {
     pub name: String,
@@ -32,593 +57,7 @@ pub struct InspectorActor {
     pub browsing_context: String,
 }
 
-#[derive(Serialize)]
-struct GetHighlighterReply {
-    highligter: HighlighterMsg, // sic.
-    from: String,
-}
-
-#[derive(Serialize)]
-struct HighlighterMsg {
-    actor: String,
-}
-
-struct HighlighterActor {
-    name: String,
-}
-
-pub struct NodeActor {
-    pub name: String,
-    script_chan: IpcSender<DevtoolScriptControlMsg>,
-    pipeline: PipelineId,
-}
-
-#[derive(Serialize)]
-struct ShowBoxModelReply {
-    from: String,
-}
-
-#[derive(Serialize)]
-struct HideBoxModelReply {
-    from: String,
-}
-
-impl Actor for HighlighterActor {
-    fn name(&self) -> String {
-        self.name.clone()
-    }
-
-    fn handle_message(
-        &self,
-        _registry: &ActorRegistry,
-        msg_type: &str,
-        _msg: &Map<String, Value>,
-        stream: &mut TcpStream,
-        _id: StreamId,
-    ) -> Result<ActorMessageStatus, ()> {
-        Ok(match msg_type {
-            "showBoxModel" => {
-                let msg = ShowBoxModelReply { from: self.name() };
-                let _ = stream.write_json_packet(&msg);
-                ActorMessageStatus::Processed
-            },
-
-            "hideBoxModel" => {
-                let msg = HideBoxModelReply { from: self.name() };
-                let _ = stream.write_json_packet(&msg);
-                ActorMessageStatus::Processed
-            },
-
-            _ => ActorMessageStatus::Ignored,
-        })
-    }
-}
-
-#[derive(Serialize)]
-struct ModifyAttributeReply {
-    from: String,
-}
-
-impl Actor for NodeActor {
-    fn name(&self) -> String {
-        self.name.clone()
-    }
-
-    fn handle_message(
-        &self,
-        registry: &ActorRegistry,
-        msg_type: &str,
-        msg: &Map<String, Value>,
-        stream: &mut TcpStream,
-        _id: StreamId,
-    ) -> Result<ActorMessageStatus, ()> {
-        Ok(match msg_type {
-            "modifyAttributes" => {
-                let target = msg.get("to").unwrap().as_str().unwrap();
-                let mods = msg.get("modifications").unwrap().as_array().unwrap();
-                let modifications = mods
-                    .iter()
-                    .map(|json_mod| {
-                        serde_json::from_str(&serde_json::to_string(json_mod).unwrap()).unwrap()
-                    })
-                    .collect();
-
-                self.script_chan
-                    .send(ModifyAttribute(
-                        self.pipeline,
-                        registry.actor_to_script(target.to_owned()),
-                        modifications,
-                    ))
-                    .unwrap();
-                let reply = ModifyAttributeReply { from: self.name() };
-                let _ = stream.write_json_packet(&reply);
-                ActorMessageStatus::Processed
-            },
-
-            _ => ActorMessageStatus::Ignored,
-        })
-    }
-}
-
-#[derive(Serialize)]
-struct GetWalkerReply {
-    from: String,
-    walker: WalkerMsg,
-}
-
-#[derive(Serialize)]
-struct WalkerMsg {
-    actor: String,
-    root: NodeActorMsg,
-}
-
-#[derive(Serialize, Clone)]
-struct AttrMsg {
-    namespace: String,
-    name: String,
-    value: String,
-}
-
-#[derive(Serialize, Clone)]
-struct NodeActorMsg {
-    actor: String,
-    baseURI: String,
-    parent: String,
-    nodeType: u16,
-    namespaceURI: String,
-    nodeName: String,
-    numChildren: usize,
-
-    name: String,
-    publicId: String,
-    systemId: String,
-
-    attrs: Vec<AttrMsg>,
-
-    pseudoClassLocks: Vec<String>,
-
-    isDisplayed: bool,
-
-    hasEventListeners: bool,
-
-    isDocumentElement: bool,
-
-    shortValue: String,
-    incompleteValue: bool,
-}
-
-/*{"walker":{"actor":"server1.conn0.process5//domwalker49","root":{"actor":"server1.conn0.process5//domnode48","baseURI":"https://servo.org/","nodeType":9,"nodeName":"#document","nodeValue":null,"displayName":"#document","numChildren":2,"displayType":null,"isScrollable":false,"isTopLevelDocument":true,"causesOverflow":false,"containerType":null,"isMarkerPseudoElement":false,"isBeforePseudoElement":false,"isAfterPseudoElement":false,"isAnonymous":false,"isNativeAnonymous":false,"isShadowRoot":false,"shadowRootMode":null,"isDirectShadowHostChild":null,"mutationBreakpoints":{"subtree":false,"removal":false,"attribute":false},"isDisplayed":true,"isInHTMLDocument":null,"traits":{}},"traits":{}},"from":"server1.conn0.process5//inspectorActor4"}*/
-
-trait NodeInfoToProtocol {
-    fn encode(
-        self,
-        actors: &ActorRegistry,
-        display: bool,
-        script_chan: IpcSender<DevtoolScriptControlMsg>,
-        pipeline: PipelineId,
-    ) -> NodeActorMsg;
-}
-
-impl NodeInfoToProtocol for NodeInfo {
-    fn encode(
-        self,
-        actors: &ActorRegistry,
-        display: bool,
-        script_chan: IpcSender<DevtoolScriptControlMsg>,
-        pipeline: PipelineId,
-    ) -> NodeActorMsg {
-        let actor_name = if !actors.script_actor_registered(self.unique_id.clone()) {
-            let name = actors.new_name("node");
-            let node_actor = NodeActor {
-                name: name.clone(),
-                script_chan,
-                pipeline,
-            };
-            actors.register_script_actor(self.unique_id, name.clone());
-            actors.register_later(Box::new(node_actor));
-            name
-        } else {
-            actors.script_to_actor(self.unique_id)
-        };
-
-        NodeActorMsg {
-            actor: actor_name,
-            baseURI: self.base_uri,
-            parent: actors.script_to_actor(self.parent.clone()),
-            nodeType: self.node_type,
-            namespaceURI: self.namespace_uri,
-            nodeName: self.node_name,
-            numChildren: self.num_children,
-
-            name: self.name,
-            publicId: self.public_id,
-            systemId: self.system_id,
-
-            attrs: self
-                .attrs
-                .into_iter()
-                .map(|attr| AttrMsg {
-                    namespace: attr.namespace,
-                    name: attr.name,
-                    value: attr.value,
-                })
-                .collect(),
-
-            pseudoClassLocks: vec![], //TODO get this data from script
-
-            isDisplayed: display,
-
-            hasEventListeners: false, //TODO get this data from script
-
-            isDocumentElement: self.is_document_element,
-
-            shortValue: self.short_value,
-            incompleteValue: self.incomplete_value,
-        }
-    }
-}
-
 // TODO: Create CSS Propperties actor
-
-struct WalkerActor {
-    name: String,
-    script_chan: IpcSender<DevtoolScriptControlMsg>,
-    pipeline: PipelineId,
-    root_node: NodeActorMsg,
-}
-
-#[derive(Serialize)]
-struct QuerySelectorReply {
-    from: String,
-}
-
-#[derive(Serialize)]
-struct DocumentElementReply {
-    from: String,
-    node: NodeActorMsg,
-}
-
-#[derive(Serialize)]
-struct ClearPseudoclassesReply {
-    from: String,
-}
-
-#[derive(Serialize)]
-struct ChildrenReply {
-    hasFirst: bool,
-    hasLast: bool,
-    nodes: Vec<NodeActorMsg>,
-    from: String,
-}
-
-#[derive(Serialize)]
-struct WatchRootNodeReply {
-    #[serde(rename = "type")]
-    type_: String,
-    from: String,
-    node: NodeActorMsg,
-}
-
-impl Actor for WalkerActor {
-    fn name(&self) -> String {
-        self.name.clone()
-    }
-
-    fn handle_message(
-        &self,
-        registry: &ActorRegistry,
-        msg_type: &str,
-        msg: &Map<String, Value>,
-        stream: &mut TcpStream,
-        _id: StreamId,
-    ) -> Result<ActorMessageStatus, ()> {
-        Ok(match msg_type {
-            "querySelector" => {
-                let msg = QuerySelectorReply { from: self.name() };
-                let _ = stream.write_json_packet(&msg);
-                ActorMessageStatus::Processed
-            },
-
-            "documentElement" => {
-                let (tx, rx) = ipc::channel().unwrap();
-                self.script_chan
-                    .send(GetDocumentElement(self.pipeline, tx))
-                    .unwrap();
-                let doc_elem_info = rx.recv().unwrap().ok_or(())?;
-                let node =
-                    doc_elem_info.encode(registry, true, self.script_chan.clone(), self.pipeline);
-
-                let msg = DocumentElementReply {
-                    from: self.name(),
-                    node,
-                };
-                let _ = stream.write_json_packet(&msg);
-                ActorMessageStatus::Processed
-            },
-
-            "clearPseudoClassLocks" => {
-                let msg = ClearPseudoclassesReply { from: self.name() };
-                let _ = stream.write_json_packet(&msg);
-                ActorMessageStatus::Processed
-            },
-
-            "children" => {
-                let target = msg.get("node").unwrap().as_str().unwrap();
-                let (tx, rx) = ipc::channel().unwrap();
-                self.script_chan
-                    .send(GetChildren(
-                        self.pipeline,
-                        registry.actor_to_script(target.to_owned()),
-                        tx,
-                    ))
-                    .unwrap();
-                let children = rx.recv().unwrap().ok_or(())?;
-
-                let msg = ChildrenReply {
-                    hasFirst: true,
-                    hasLast: true,
-                    nodes: children
-                        .into_iter()
-                        .map(|child| {
-                            child.encode(registry, true, self.script_chan.clone(), self.pipeline)
-                        })
-                        .collect(),
-                    from: self.name(),
-                };
-                let _ = stream.write_json_packet(&msg);
-                ActorMessageStatus::Processed
-            },
-
-            "watchRootNode" => {
-                let _ = stream.write_json_packet(&WatchRootNodeReply {
-                    type_: "root-available".into(),
-                    from: self.name(),
-                    node: self.root_node.clone(),
-                });
-
-                let _ = stream.write_json_packet(&EmptyReplyMsg { from: self.name() });
-
-                ActorMessageStatus::Processed
-            },
-
-            _ => ActorMessageStatus::Ignored,
-        })
-    }
-}
-
-#[derive(Serialize)]
-struct GetPageStyleReply {
-    from: String,
-    pageStyle: PageStyleMsg,
-}
-
-#[derive(Serialize)]
-struct PageStyleMsg {
-    actor: String,
-}
-
-struct PageStyleActor {
-    name: String,
-    script_chan: IpcSender<DevtoolScriptControlMsg>,
-    pipeline: PipelineId,
-}
-
-#[derive(Serialize)]
-struct GetAppliedReply {
-    entries: Vec<AppliedEntry>,
-    rules: Vec<AppliedRule>,
-    sheets: Vec<AppliedSheet>,
-    from: String,
-}
-
-#[derive(Serialize)]
-struct GetComputedReply {
-    computed: Vec<u32>, //XXX all css props
-    from: String,
-}
-
-#[derive(Serialize)]
-struct AppliedEntry {
-    rule: String,
-    pseudoElement: Value,
-    isSystem: bool,
-    matchedSelectors: Vec<String>,
-}
-
-#[derive(Serialize)]
-struct AppliedRule {
-    actor: String,
-    #[serde(rename = "type")]
-    type_: String,
-    href: String,
-    cssText: String,
-    line: u32,
-    column: u32,
-    parentStyleSheet: String,
-}
-
-#[derive(Serialize)]
-struct AppliedSheet {
-    actor: String,
-    href: String,
-    nodeHref: String,
-    disabled: bool,
-    title: String,
-    system: bool,
-    styleSheetIndex: isize,
-    ruleCount: usize,
-}
-
-#[derive(Serialize)]
-struct GetLayoutReply {
-    from: String,
-
-    display: String,
-    position: String,
-    #[serde(rename = "z-index")]
-    zIndex: String,
-    #[serde(rename = "box-sizing")]
-    boxSizing: String,
-
-    // Would be nice to use a proper struct, blocked by
-    // https://github.com/serde-rs/serde/issues/43
-    autoMargins: serde_json::value::Value,
-    #[serde(rename = "margin-top")]
-    marginTop: String,
-    #[serde(rename = "margin-right")]
-    marginRight: String,
-    #[serde(rename = "margin-bottom")]
-    marginBottom: String,
-    #[serde(rename = "margin-left")]
-    marginLeft: String,
-
-    #[serde(rename = "border-top-width")]
-    borderTopWidth: String,
-    #[serde(rename = "border-right-width")]
-    borderRightWidth: String,
-    #[serde(rename = "border-bottom-width")]
-    borderBottomWidth: String,
-    #[serde(rename = "border-left-width")]
-    borderLeftWidth: String,
-
-    #[serde(rename = "padding-top")]
-    paddingTop: String,
-    #[serde(rename = "padding-right")]
-    paddingRight: String,
-    #[serde(rename = "padding-bottom")]
-    paddingBottom: String,
-    #[serde(rename = "padding-left")]
-    paddingLeft: String,
-
-    width: f32,
-    height: f32,
-}
-
-impl Actor for PageStyleActor {
-    fn name(&self) -> String {
-        self.name.clone()
-    }
-
-    fn handle_message(
-        &self,
-        registry: &ActorRegistry,
-        msg_type: &str,
-        msg: &Map<String, Value>,
-        stream: &mut TcpStream,
-        _id: StreamId,
-    ) -> Result<ActorMessageStatus, ()> {
-        Ok(match msg_type {
-            "getApplied" => {
-                //TODO: query script for relevant applied styles to node (msg.node)
-                let msg = GetAppliedReply {
-                    entries: vec![],
-                    rules: vec![],
-                    sheets: vec![],
-                    from: self.name(),
-                };
-                let _ = stream.write_json_packet(&msg);
-                ActorMessageStatus::Processed
-            },
-
-            "getComputed" => {
-                //TODO: query script for relevant computed styles on node (msg.node)
-                let msg = GetComputedReply {
-                    computed: vec![],
-                    from: self.name(),
-                };
-                let _ = stream.write_json_packet(&msg);
-                ActorMessageStatus::Processed
-            },
-
-            //TODO: query script for box layout properties of node (msg.node)
-            "getLayout" => {
-                let target = msg.get("node").unwrap().as_str().unwrap();
-                let (tx, rx) = ipc::channel().unwrap();
-                self.script_chan
-                    .send(GetLayout(
-                        self.pipeline,
-                        registry.actor_to_script(target.to_owned()),
-                        tx,
-                    ))
-                    .unwrap();
-                let ComputedNodeLayout {
-                    display,
-                    position,
-                    z_index: zIndex,
-                    box_sizing: boxSizing,
-                    auto_margins: autoMargins,
-                    margin_top: marginTop,
-                    margin_right: marginRight,
-                    margin_bottom: marginBottom,
-                    margin_left: marginLeft,
-                    border_top_width: borderTopWidth,
-                    border_right_width: borderRightWidth,
-                    border_bottom_width: borderBottomWidth,
-                    border_left_width: borderLeftWidth,
-                    padding_top: paddingTop,
-                    padding_right: paddingRight,
-                    padding_bottom: paddingBottom,
-                    padding_left: paddingLeft,
-                    width,
-                    height,
-                } = rx.recv().unwrap().ok_or(())?;
-
-                let auto_margins = msg
-                    .get("autoMargins")
-                    .and_then(Value::as_bool)
-                    .unwrap_or(false);
-
-                // http://mxr.mozilla.org/mozilla-central/source/toolkit/devtools/server/actors/styles.js
-                let msg = GetLayoutReply {
-                    from: self.name(),
-                    display,
-                    position,
-                    zIndex,
-                    boxSizing,
-                    autoMargins: if auto_margins {
-                        let mut m = Map::new();
-                        let auto = serde_json::value::Value::String("auto".to_owned());
-                        if autoMargins.top {
-                            m.insert("top".to_owned(), auto.clone());
-                        }
-                        if autoMargins.right {
-                            m.insert("right".to_owned(), auto.clone());
-                        }
-                        if autoMargins.bottom {
-                            m.insert("bottom".to_owned(), auto.clone());
-                        }
-                        if autoMargins.left {
-                            m.insert("left".to_owned(), auto);
-                        }
-                        serde_json::value::Value::Object(m)
-                    } else {
-                        serde_json::value::Value::Null
-                    },
-                    marginTop,
-                    marginRight,
-                    marginBottom,
-                    marginLeft,
-                    borderTopWidth,
-                    borderRightWidth,
-                    borderBottomWidth,
-                    borderLeftWidth,
-                    paddingTop,
-                    paddingRight,
-                    paddingBottom,
-                    paddingLeft,
-                    width,
-                    height,
-                };
-                let msg = serde_json::to_string(&msg).unwrap();
-                let msg = serde_json::from_str::<Value>(&msg).unwrap();
-                let _ = stream.write_json_packet(&msg);
-                ActorMessageStatus::Processed
-            },
-
-            _ => ActorMessageStatus::Ignored,
-        })
-    }
-}
 
 impl Actor for InspectorActor {
     fn name(&self) -> String {
@@ -680,7 +119,7 @@ impl Actor for InspectorActor {
 
                 let msg = GetPageStyleReply {
                     from: self.name(),
-                    pageStyle: PageStyleMsg {
+                    page_style: PageStyleMsg {
                         actor: self.page_style.borrow().clone().unwrap(),
                     },
                 };
